Move health check rate limiting test to OrchardCore.HealthChecks test module
